app/requirements.txt
discord
python-dotenv
psycopg2